Implements persisting room tables

+++ persistence/src/main/java/com/bopera/pointofsales/persistence/repository/RoomTablesRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RoomTablesRepository extends BaseRepository<RoomTableEntity, Integer> {
+public interface RoomTablesRepository extends BaseRepository<RoomTableEntity, Long> {
 
     @Query(value = "SELECT * FROM tables t WHERE t.room_id = :roomId", nativeQuery = true)
-    List<RoomTableEntity> findByRoomId(Integer roomId);
+    List<RoomTableEntity> findByRoomId(Long roomId);
 }
